Fixed some small bugs in the grammar and the example.
I also added the following to Node.java:

public Object ext;   // the extension object--has to be public or else visitors can't see it.

public boolean isExtended();

I still need to alter the constructors and reconstruct methods to add the extension objects and change the frontend so that the Jif tools generate Jif-extended objects.