"The Bomb"
Rewrote types and ast hierarchy.  Changed naming conventions to be less
verbose.  Made Node and descendants use a more functional style.  Most
type objects are functional too, except for ParsedClassType and friends.

Node interface files are in jltools/ast and jltools/types.  The standard
implementations of these interfaces are in jltools/ext/jl/ast and types.

Example:

"While" is the interface for a "while" statement.  It looks like this:

public interface While extends Stmt
{
    Expr cond();
    Stmt body();

    While cond(Expr cond);
    While body(Stmt body);
}

The first two methods get fields in the node.  The second two methods
set those fields by first copying the While object and then setting the
fields in the copy.  "reconstruct" is no longer public, but is still
used by "visitChildren".

To create a new "While" node, call "nf.While(position, cond, body)",
where "nf" is an instance of "NodeFactory".  The "position" argument
specifies the position in the source file of the while statement.  For
most nodes, only the parser needs to worry about creating nodes.  The
other main use is with disambiguation.

"While_c" is the default implementation of the "While" interface.  The
usual idiom for setting a field in the implementation class is:

    public While cond(Expr cond) {
	While_c n = (While_c) copy();
	n.cond = cond;
	return n;
    }

Node_c and TypeObject_c implement the copy() method.  copy is just a
wrapper around clone() which catches an exception clone() should never
throw.

All nodes have extension objects, accessible from with the ext() method.
All extension objects should have a reference to their node, so it need
not be passed to the extension.  All pass methods should be called
through the extension, as in:

	n.ext().foo()

or:

	n.ext().fooOverride()

The call is indirected through the extension object to ensure language
extensions always have a hook to override the method.  Most of the
methods should only be called from the visitor anyway, so the
n.ext().foo() bit needs to appear there.  The translate pass, though,
does not use a visitor, so you will see the indirect calls there.

All Nodes implement two methods for each pass (except translate):

    Node fooOverride_(FooVisitor v) is called from the FooVisitor
    override() method.  The visitor should give access to any other
    objects the Node needs, for instance TypeSystem, NodeFactory, or
    Context.  The default implementation of fooOverride_ (in Node_c)
    returns null.  If non-null is returned, the visitor will not visit
    children and will not call enter() and leave() for the node.

    Node foo_(FooVisitor v) is called from the FooVisitor leave()
    method.  The default implementation returns this.

The purpose of having two methods is so that the visitor does not need
to know which Nodes need to use override() and which need to use
leave(); it just calls both.

The "_" at the end of the method names is a hint that these methods
should not be called directly, but rather through the extension, which
does not have the "_" in the method name.

The passes are now as follows:

    buildTypes - constructs the types.  For classes, these are instances
    of ParsedClassType.  For methods, MethodInstance.  Similarly,
    FieldInstance, ConstructorInstance, etc.  All Expr nodes after this
    phase should have type <unknown>.  This fixes some null pointer
    errors when type checking fails and a type is not assigned.

    disambiguateTypes - all TypeNodes (and labels in Jif) should be
    disambiguated after this phase.  There are now several kinds of
    TypeNodes.  CanonicalTypeNodes contain canonical types (canonical
    just mean that the type is completely disambiguated).
    ArrayTypeNodes contain another a base TypeNode, usually
    non-canonical.  Jif adds LabeledTypeNode, and InstTypeNode.  Giving
    the TypeNodes some structure allows all disambiguation to be done on
    the AST side and cleans up some wierd cases in Jif, as well.
    TypeSystem.checkAndResolveType and LocalContext.getType have gone
    away.

    disambiguate - this disambiguates all nodes other than type nodes.

    foldConstants - this just folds constants and does some algebraic
    simplification

    typeCheck - this type checks.  Expr.type() should return a canonical
    type after this phase.

    exceptionCheck - this does exception checking

    translate - this writes the ast to a CodeWriter

The disambiguateTypes pass may be run more than once.  This can happen
when a class declaration is visited before the type objects in its super
class are disambiguated.   If this happens, disambiguation of the class
is delayed and the pass is repeated after the super class is
disambiguated.  This solution is less efficient than I would like, but
it works.

Annotations are gone, but live on in fields of the various classes
they were annotating.

Moved most of the work to enter the scope of a class or method from
Context into the class or method decl node.

Fixed ImportTable to not import classes in other packages that are not
declared public.

Merged ClassResolver and TypeContext into one.  Both are maps from names
to types, and they share a lot of code.

Rewrote LocalContext (now called Context) to be easier to extend.

TypeObject is the super type of Type, Package, FieldInstance, MethodInstance,
etc.  All TypeObjects must be constructed via the type system, e.g.,

	ArrayType arrayOf(Type t)
	PrimitiveType Int()
	TopLevelClass topLevelClassType()
	MethodInstance methodInstance()

Packages are no longer Types.  ClassType has been split into four classes:
TopLevelClassType, MemberClassType, LocalClassType, and AnonymousClassType.
Why?  Mainly to get rid of the need to create bogus fully-qualified names for
anonymous and local classes.

Jif changes:

I merged Components and Labels into one hierarchy.  JoinLabel is a label
which is a join of other labels.  leq, meet, and join go through the type
system, so new label types can be added there.  Cleaned up label
disambiguation.  Removed JifUtil.checkCall.  Simplified call and method
declaration checking so we don't have to create a dummy environment in which to
interpret types.  (see ArgLabel and DynamicArgLabel).

What's still broken:

Do enter/leaveScope need to go through the ext?  So far this hasn't been
necessary, but it may in the future.

I'm not entirely happy with how the pass methods in a node are split into an
override and a leave method.  The problem is that I don't really want the
subclasses to know which to override when they just want to do a small bit of
work (e.g., adding a constraint) after visiting the node.
Maybe there should be a method that is always called after both the leave and
override methods.

Jif: We need a pc for FieldDecl.init()

Jif: field labels aren't right.  They should not be variables.

Jif: is the "this" label done correctly?