The Parser Cleanup
The PolyJ parser, and to a lesser extent the java12 (or JL) parser,
had grown a "parser code" section that was far too big.  Although I've
been trimming the PolyJ parser all summer, this is the final cut to
the bare minimum of code necessary in that section.  The size of the
CUP file has been reduced by about 1000 lines since I started.
The reduced size should (hopefully) make the parsers easier to
debug and maintain.

The code that all of the parsers have in common has been moved to
jltools.parse.Util.  Since all the parsers seem to use the old
Grm.Declarator as-is, I've made it jltools.parse.VarDeclarator (the
name change is to avoid conflicts with jltools.ast.Declarator).
All static inner classes have been made top-level classes in their
own files.  Methods particular to an individual parser have been
moved to that parser's own override of Util (e.g., PolyJUtil).

To use the Util class, an extension-compiler needs to provide
its own override of Util, its own implementation of ParserWrapper,
and then construct the necessary objects in its parser's constructor.
See either the JL parser or the PolyJ parser for an example.