Changed extension nodes.  Each Node has a Del (delegate) through which methods should be invoked.  The Del has a pointer to zero or more Exts. The method which returns the Ext is typically named after the language. So, in jif/split, to access the Jif extension:
    ((SplitDel) n.del()).jif()

and to access the Split extension:

    ((SplitDel) n.del()).split()

I know this is cumbersome, but it's more efficient than just having
each node have a Map of Exts.  For most of the code, the casts are just
put in a method, so you can do something like the following to get and
set the extension for your language:

    Jif ext = jifExt(n);
    JifExprExt ext = (JifExprExt) jifExt(n);
    Node m = jifExt(n, newExt);

I also renamed all the foo_() methods to just foo().  The compiler should
issue a warning if you try to call foo() on a node.