Modified the dataflow infrastructure to permit finer-grained dataflow analyses.
In particular, dataflow can now be flow sensitive, that is, the Items that are passed on the true and false edges leaving a boolean condition may be different. For example, in the code "if (o != null) o.foo();", a not-null analysis can now determine that at the call o.foo(), o is not null.

This change was accomplished primarily by altering the class polyglot.visit.FlowGraph, so that Items are associated with edges, rather than nodes in the control flow graph, and modifying polyglot.visit.Dataflow to permit subclasses to have the appropriate information and infrastructure to take advantage of this change.

Also modified Jif's NotNullChecker to take advantage of this change.

Other bugs fixed during this change:
- CFGBuilder was incorrectly flowing exception throws to inappropriate catch blocks.
- a default implementation of exceptionCheck was provided in Node_c, that calls Thrower.throwsTypes, to try and reduce the inconsistencies between these two methods.
- The expression Throw can now throw a NullPointerException (which is what "throw null;" will produce)
- Changed the visitAssignCFG method for ArrayAccess_c and Field_c, so that the ArrayAccess and Field nodes respectively are always visited, to ensure that some node in the CFG is throwing the NullPointerException, and ArrayIndexOutOfBoundsExceptions that these expressions can generate.

Also modified:
- Jif's NotNullChecker now checks throw and array access statements.