Added field to Field AST node, recording if the target is implicit, e.g. if the source code said "f", which got disambiguated to "this.f", then the target is implicit, vs. if the source code said "this.f" then the target is not implicit.
This distinction is important for compile time constants. Still need to implement a check that ensures that the assertion that the target is implicit is consistent with the context, in case the field gets moved/copied to another context after disambiguation. Also need to modify initialization checking to make use of this new field.