Bug fixes:
Fixed null pointer bugs.

Fixed flow operation to account for variables going out of
scope.


Tweaks to improve optimization:

Improved confluence operation -- to intersect a forest with
another, remove a link between two nodes if they don't
appear in the same tree in the other forest.

When killing a variable, splice it out of the tree rather
than split the tree apart.