Goals now have a method to return their distance from the goal.  This is typically just 0 or 1, but some goals such as Disambiguated return the number of ambiguous nodes in the AST.  The scheduler will abort the compilation if a goal is attempted twice with no progress (that is reduction in distance) made between attempts.
Modified AmbiguityRemover to avoid disambiguating method bodies
or member signatures.  This is taken advantage of by
SignaturesResolved and SupertypesResolved to avoid some unnecessary work.

Added command-line option to compile only files given on the command-line.
Other files may be loaded on demand and parsed, but might not be compiled
to java.