General purpose classes for rewriting between language extension ASTs. Generalized from JifToJavaRewriter.  Extensions may utilize this by adding a ToExtFactory to their NodeFactory and extending ToExt classes as necessary.
The ExtensionRewriter is a visitor that finds the appropriate ToExt extension
object for the target extension using the getToExt() method defined in
ExtensionInfo. The default is to retrieve the first available ToExt object, but
this could be overridden to support multiple ToExt targets.

Implementing extensions will probably need to implement an "output"
ExtensionInfo that subclasses the target extension.  At the very least, this
extension will need a schedluer that skips the default "Parsed" goal.