Fix a bug in the implementation of moreSpecificImpl(ProcedureInstance).
Although the JLS 3rd edition doesn't say so, it seems that we need to ensure that t1 (the container of p1) can be converted to t2 (the container of p2) by method invocation conversion. At least, this is what happens in JLS 2nd edition, and it handles the case where p1 overrides p2 (since otherwise, both p1 and p2 would fit the definition for maximally specific).