Fix annotation initialization goals. Ensure all classes are ready to be rewritten before starting.
For source files, retained annotations are not set in the type until type checking.  Now the class initializer throws the correct MissingDependencyException.

Rewriting from one JL5 to JL cannot be interrupted (for instance, due to a MDE) in a recoverable way.  Ensure that all jobs are ready for rewriting before beginning to rewrite.

Also tweaked visibility of annotation ext impl methods and the EndGoal constructor for extensibility.