Fix algorithm for finding lub of an anonymous class and another type.
When generating the Erased Super Type set, compute the erasures of the super types of each element rather than the super types of the erasure. This is necessary because ParsedClassTypes should be treated as Raw types in this context but do not have an erase method.

Also consider a raw type to be an invocation of a generic type when solving for the correct substitution.