More complete support for rewriting JL5 to JL5.
Some classes were implicitly converting from JL5 to JL by using a ToExt
object that was oblivious to JL5 features.  This makes those situations
more explicit and introduces a new rewriter for JL5 to JL5 that handles
conversion of types with paramaters.

NB: This last part in particular is still experimental, so buyer beware ;)