Modified translation of JL5New and TypeNodes to allow ambiguous types.
In general, it has been safe to translate ambiguous nodes.  This allows
extensions to use the quasiquoter to emit code that isn't going to be
typechecked by an output extension.

Some recent fixes for translating fully qualified names break when they
try to access the type of a node that hasn't been typechecked.  This commit
fixes that for a couple of places, but it is probably worth looking for
any code that accesses type() methods during translation, and guard them
with a safe alternative.