Restructured the construction of node and extension objects.
The constructor of each node class should take an extension to be attached to it.  The existing constructors are marked as deprecated.

A node factory should create all extension objects up to the factory's extension level.  The constructor of extension class at the factory's extension level can take appropriate arguments representing additional fields (data) in the extension.  The resulting extension object chain is then passed to the superlanguage factory for the remaining object creation.