Fixed the definition of inheritance.  Added relevant test cases.
Class C does not inherit a method defined in class A if there is a class
B such that C <: B <: A such that C != B != A and B also defines a
method of the same signature, i.e., overriding A's method.

Modified pth's grammar to handle empty list of tests gracefully.