Remove caching for RawClass methods and fields
RawClass previously cached the methods of the underlying SubstClass,
but this could cause problems if the base class of the SubstClass
changed. Specifically, changes (e.g., an added method) to the
base class would never propagate to the RawClass, and in fact the
compiler would crash inside CachingTransformingList due to the
mismatch.

The underlying base class can change during, for example,
desugar transformations.

The performance impact of this change should not be significant.

Test: testsjl5/pthScript