Rethrow missing deps within resolveCaseLabels
This resolves a subtle null pointer exception that could
occur when resolveCaseLabel threw a missing dependency
exception, thereby leaving a case expression with a null type
while type checking the enclosing switch statement.

By rethrowing the missing dependency exception, we ensure that the
enclosing switch statement does not touch the case expression
until the missing dependency is satisfied.

This came up when trying to compile the JDK with Polyglot.