Fix subtle bug in method resolution
(1) Interfaces without a super-interfaces implicitly declare abstract
versions of the methods of Object. The `.declaration()` of each of these
still points to the corresponding method in Object.

(2) JL5RawSubst#substMethod uses `.declaration()` as the starting point
for the substituted method.

The two facts above led to some raw-substituted methods having incorrect
flags, which in turn caused problems for method resolution.