Fix subtle bug in ArrayType#base(Type)
The base type of an ArrayType could be changed by calling
ArrayType#base(...). But the array type returned would
still have the fields and methods of the previous array
type. Thus, the "container" of each field and method
would still point to the old base type, and this
inconsistency would result in occasional failures
when searching for fields in an array type.

This bug appears when trying to compile JDK 7.