Fix compiler crash related to anonymous enum class
If an anonymous enum class referred (within its body) to another
enum constant from the same enum type, then the pretty printer
would attempt to qualify that field access with a canonical type
node corresponding to an anonymous class---which, of course, has
no textual representation. This one-line change ensures that the
field access target is implicit, and thus not emitted during
pretty-printing.