Implement a method filter flag
This allows the user to specify a file containing a list of
regular expressions matching method signatures. For each
method matched, the body of that method is replaced with
`throw new RuntimeException(...)` prior to type checking.

This is useful for compiling the JDK, for example, where some methods
seem to rely on type checking behavior that is peculiar to
javac, but not specified in the JLS.