Fix definite assignment status after ctor calls
The AssignmentChecker makes a list of constructors that
call another constructor within the body, and uses this
information as part of ensuring that non-static final fields
are always initialized by the end of every constructor.

However, the assignment status of non-static final
fields remained unchanged on the control flow graph edge
leaving the constructor call, meaning that we could incorrectly
report errors if a non-static final field is used in the body of
a constructor.

Here's an example of something that previously failed to compile.

class A {
    private final int i;

    A() {
        i = 0;
    }

    A(int n) {
        this();
        int j = i; // Error: "i might not have been initialized."
    }
}

Test: testsjl5